fix for labeling of area plots and opt gaps plots
previously, all the area plots (i.e. plots of time vs number of qubits', with objective function evolution shown in color has incorrect  labelling on colorbar, wherein it always said avg. approximation ratio.) With this modification, it is labelled with a string that correctly reflects the choice of the objective function